refactor(net): move node record constants to network-peers crate
